// workshop/services/sample/src/generated/types.gen.ts
export type ResourceStatus = 'active' | 'inactive' | 'pending' | 'archived';

/**
 * Status of a resource
 */
export const ResourceStatus = {
    ACTIVE: 'active',
    INACTIVE: 'inactive',
    PENDING: 'pending',
    ARCHIVED: 'archived'
} as const;